LayoutTests/http/tests/security/load-image-after-redirection-2-expected.txt

CONSOLE MESSAGE: Cross-origin redirection to http://127.0.0.1:8000/security/resources/allow-if-origin.php denied by Cross-Origin Resource Sharing policy: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
 

PASS Loading an image in no-cors mode after a CORS invalid redirection is ok 
PASS Loading an image in cors mode after a CORS invalid redirection is stopped 
